bdf_vectorized3:  - adding RADSET, SESET, CAABSF, PAABSF, QVECT, RANDPS  - more add/clear methods  - more work on MATORT
op2:
 - better shock support

+++ pyNastran/dev/bdf_vectorized3/cards/elements/fast.py
@@ -108,7 +108,7 @@ class CFAST(Element):
         nodes = np.zeros((ncards, 2), dtype=idtype)
 
         fast_type = np.zeros(ncards, dtype='|U4')
-        ids = np.zeros((ncards, 2), dtype='int32')
+        ids = np.zeros((ncards, 2), dtype=idtype)
         fastener_node = np.zeros(ncards, dtype='int32')
         fastener_xyz = np.zeros((ncards, 3), dtype='float64')
         for icard, card in enumerate(self.cards):
